+++ src/app/core/components/graphs/StackedAreaChart.tsx
@@ -1,43 +1,7 @@
-import React from 'react'
+import React, { ReactNode } from 'react'
 import { AreaChart, Area, XAxis, YAxis, CartesianGrid, ResponsiveContainer, Tooltip } from 'recharts'
 import { useTheme } from '@material-ui/core/styles'
 import { pathStr } from 'utils/fp'
-import { makeStyles } from '@material-ui/styles'
-import FontAwesomeIcon from 'core/components/FontAwesomeIcon'
-import { Theme } from '@material-ui/core'
-
-const useStyles = makeStyles((theme: Theme) => ({
-  customTooltip: {
-    background: '#ffffff',
-    padding: theme.spacing(1),
-    fontSize: 13,
-    minWidth: 144,
-  },
-  customTooltipHeader: {
-    borderBottom: '1px solid #d8d8d8',
-    padding: theme.spacing(1),
-  },
-  customTooltipBody: {
-    paddingTop: theme.spacing(2),
-  },
-  customTooltipEntry: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    paddingLeft: theme.spacing(1),
-    paddingRight: theme.spacing(1),
-    paddingBottom: theme.spacing(1),
-  },
-  customTooltipLabel: {
-    flexGrow: 0,
-  },
-  customTooltipName: {
-    marginLeft: theme.spacing(1)
-  },
-  customTooltipCount: {
-    flexGrow: 0,
-    paddingRight: theme.spacing(1)
-  },
-}))
 
 /*
   Usage:
@@ -88,49 +52,7 @@ interface Props<T extends string, V extends string> {
   keys: Array<AreaChartType<V>>
   responsive?: boolean
   verticalAxisLines?: boolean
-  useCustomTooltip?: boolean
-}
-
-interface CustomTooltipProps {
-  active?: boolean
-  payload?: any
-  label?: string
-  keys: any[]
-}
-
-// Should the Label and Entries be passed in as a prop?
-const CustomTooltip = ({ active, payload, label, keys }: CustomTooltipProps) => {
-  const classes = useStyles({})
-  const theme: any = useTheme()
-
-  if (active && payload) {
-    // Need data from both payload (count) & keys (icon, color)
-    const countByKey = payload.reduce((acc, value) => {
-      acc[value.name] = value.payload[value.name]
-      return acc
-    }, {})
-
-    return (
-      <div className={classes.customTooltip}>
-        <div className={classes.customTooltipHeader}>
-          Alarms @ {label}
-        </div>
-        <div className={classes.customTooltipBody}>
-          {keys.map(({ name, color, icon }) => (
-            <div key={name} className={classes.customTooltipEntry}>
-              <div className={classes.customTooltipLabel}>
-                <FontAwesomeIcon solid style={{color: pathStr(color, theme.palette)}}>{icon}</FontAwesomeIcon>
-                <span className={classes.customTooltipName}>{name}</span>
-              </div>
-              <div className={classes.customTooltipCount}>{countByKey[name]}</div>
-            </div>
-          ))}
-        </div>
-      </div>
-    )
-  }
-
-  return null
+  customTooltip?: ReactNode
 }
 
 // Todo: Instead of a set width and height, also allow for percents
@@ -142,7 +64,7 @@ function StackedAreaChart<Axis extends string, Types extends string>({
   xAxis,
   verticalAxisLines = false,
   responsive = false,
-  useCustomTooltip = false,
+  customTooltip,
 }: Props<Axis, Types>) {
   const theme: any = useTheme()
 
@@ -158,7 +80,7 @@ function StackedAreaChart<Axis extends string, Types extends string>({
         <Tooltip
           {...{
             cursor: {stroke: 'rgba(96, 96, 96, 0.5)', strokeWidth: 6},
-            ...(useCustomTooltip && { content: <CustomTooltip keys={keys} />}),
+            ...(customTooltip && { content: customTooltip }),
           }}
         />
         {keys.map(({ name, color }) => (
